@@ -14,31 +14,32 @@
     "@tonaljs/tonal": "^4.6.5",
     "audiokeys": "^0.1.1",
     "chart.js": "^3.9.1",
-    "vue": "^2.7.14",
-    "vue-router": "^3.5.4",
-    "vuex": "^3.6.2",
-    "vue-chartjs": "^4.1.2",
-    "webmidi": "^3.0.10",
-    "tweakpane": "^3.1.10",
-    "three": "^0.157",
-    "tone": "^14.7.77",
+    "js-yaml": "^4.1.0",
+    "meyda": "^5.6.2",
     "remixicon": "^3.5.0",
     "ringbuf.js": "^0.3.6",
+    "three": "^0.157",
+    "tone": "^14.7.77",
+    "tweakpane": "^3.1.10",
+    "vexflow": "4.2.3",
+    "vue": "^2.7.14",
+    "vue-chartjs": "^4.1.2",
     "vue-js-modal": "^2.0.1",
     "vue-js-toggle-button": "^1.3.3",
     "vue-material": "^1.0.0-beta-16",
+    "vue-router": "^3.5.4",
     "vue-simple-search-dropdown": "^1.0.1",
     "vue-toasted": "^1.1.28",
     "vue2-touch-events": "^3.2.3",
-    "js-yaml": "^4.1.0",
-    "meyda": "^5.6.2",
-    "vexflow": "4.2.3"
+    "vuex": "^3.6.2",
+    "webmidi": "^3.0.10"
   },
   "devDependencies": {
     "@vitejs/plugin-legacy": "^2.0.0",
     "@vitejs/plugin-vue2": "^1.1.2",
-    "eslint": "^8.5.0",
-    "eslint-plugin-vue": "^9.0.0",
+    "eslint": "^8.52.0",
+    "eslint-config-google": "^0.14.0",
+    "eslint-plugin-vue": "^9.18.1",
     "terser": "^5.14.2",
     "vite": "^3.0.2"
   }
