@@ -4,10 +4,10 @@
  * so don't do any heavy processing here.
  * This hook is only invoked if in config.yaml you have set:
  * interactionMode.audioMode: true
- * 
- * @param {Array<Float32Array>} buffer - An array of Float32Array representing 
+ *
+ * @param {Array<Float32Array>} buffer - An array of Float32Array representing
  * audio data.Each Float32Array corresponds to a channel of audio data.
  */
 export function processAudioBuffer(buffer) {
-    // Put your code here
+  // Put your code here
 }
