@@ -3,43 +3,48 @@ import {
     bachDuetNoteToEuterpeNote,
     euterpeTickToBachDuetTick,
     bachDuetInference,
-    resetBachDuetState
-    } from './bachDuetUtils.js';
+} from './bachDuetUtils.js';
 
+/**
+ * Processes a clock event and returns a message with the agent's output notes.
+ * @param {Object} content - The clock event content.
+ * @return {Object} - The message with the agent's output notes.
+ */
 export function processClockEvent(content) {
-    
-    let numUserNotes = content.humanQuantizedInput.length;
+    const numUserNotes = content.humanQuantizedInput.length;
     if (numUserNotes > 1) {
-        console.error("BachDuet can only handle monophonic input. Make sure to set polyphony.input:1 in config.yaml");
+        console.error('BachDuet can only handle monophonic input. ' +
+        'Make sure to set polyphony.input:1 in config.yaml');
     }
     // If there is no user note, BachDuet expects the a 'rest' token/note
-    let userInput_bd;
+    let userInputBD;
     if (numUserNotes == 0) {
-        userInput_bd = self.restNote;
+        userInputBD = self.restNote;
     } else {
-        userInput_bd = euterpeNoteToBachDuetNote(content.humanQuantizedInput[0]);
+        userInputBD = euterpeNoteToBachDuetNote(content.humanQuantizedInput[0]);
     }
     // Convert the tick from Euterpe's format to BachDuet's format
-    let tick_bd = euterpeTickToBachDuetTick(content.tick);
+    const tickBD = euterpeTickToBachDuetTick(content.tick);
 
     // Prepare the input tensors for the neural network
-    let midiInput_t = tf.tensor2d([[self.lastBachDuetNote.midiArticInd, userInput_bd.midiArticInd]]);
-    let cpcInput_t = tf.tensor2d([[self.lastBachDuetNote.cpc, userInput_bd.cpc]]);
-    let tick_t = tf.tensor2d([[tick_bd]]);
+    const midiInputTensor = tf.tensor2d([[self.lastBachDuetNote.midiArticInd,
+        userInputBD.midiArticInd]]);
+    const cpcInputTensor = tf.tensor2d([[self.lastBachDuetNote.cpc, userInputBD.cpc]]);
+    const tickTensor = tf.tensor2d([[tickBD]]);
 
     // Neural network inference
-    let currentBachDuetNote = bachDuetInference(midiInput_t, cpcInput_t, tick_t);
-    
+    const currentBachDuetNote = bachDuetInference(midiInputTensor, cpcInputTensor, tickTensor);
+
     // Convert the bachDuet's note output to Euterpe's note format
-    let agentOutputNoteList = bachDuetNoteToEuterpeNote(currentBachDuetNote);
-    
+    const agentOutputNoteList = bachDuetNoteToEuterpeNote(currentBachDuetNote);
+
     // Update the last note (it'll be used in the next tick)
     self.lastBachDuetNote = currentBachDuetNote;
 
     // Add your messages to Euterpe here
-    let message = {
-        [self.messageType.NOTE_LIST]: agentOutputNoteList
-    }
+    const message = {
+        [self.messageType.NOTE_LIST]: agentOutputNoteList,
+    };
 
     return message;
 }
