@@ -3,68 +3,68 @@
         <span class="label">{{ internalValue }}</span>
         <div class="range">
             <input class="horizontal slider" type="range" :min="min" :max="max" v-model="internalValue"
-                @change="handleSliderChange($event.target.value)" 
+                @change="handleSliderChange($event.target.value)"
                 @input="handleSliderDrag($event.target.value)"
             />
         </div>
     </div>
 </template>
-  
+
 <script>
 // Added last minute.
 // TODO: Unify style with other sliders.
 
 export default {
-    name: 'BPMSlider',
-    props: {
-        value: {
-            type: Number,
-            required: true,
-            default: 80
-        },
-        min: {
-            type: Number,
-            default: 60
-        },
-        max: {
-            type: Number,
-            default: 120
-        },
+  name: 'BPMSlider',
+  props: {
+    value: {
+      type: Number,
+      required: true,
+      default: 80,
     },
-    
-    data() {
-        return {
-            internalValue: this.value
-        }
+    min: {
+      type: Number,
+      default: 60,
+    },
+    max: {
+      type: Number,
+      default: 120,
+    },
+  },
+
+  data() {
+    return {
+      internalValue: this.value,
+    };
+  },
+
+  watch: {
+    value(newVal) {
+      this.internalValue = newVal;
+    },
+  },
+
+  mounted() {
+    this.handleSliderDrag(this.internalValue);
+  },
+
+  methods: {
+    handleSliderChange(value) {
+      // When the slider value changes, emit the 'input' event
+      this.$emit('bpmChangeEvent', Number(value));
+      this.handleSliderDrag(value);
     },
 
-    watch: {
-        value(newVal) {
-            this.internalValue = newVal;
-        },
+    handleSliderDrag(value) {
+      // Update the background image for the slider
+      const percent100 = (value - this.min) / (this.max - this.min) * 100;
+      const bg = `linear-gradient(90deg, var(--bg-value-color) 0%, var(--bg-value-color) ${percent100}%, var(--bg-range-color) ${percent100}%, var(--bg-range-color) 100%)`;
+      this.$el.querySelector('input').style.backgroundImage = bg;
     },
-
-    mounted() {
-        this.handleSliderDrag(this.internalValue);
-    },
-
-    methods: {
-        handleSliderChange(value) {
-            // When the slider value changes, emit the 'input' event
-            this.$emit('bpmChangeEvent', Number(value));
-            this.handleSliderDrag(value);
-        },
-
-        handleSliderDrag(value) {
-            // Update the background image for the slider
-            let percent100 = (value - this.min) / (this.max - this.min) * 100;
-            let bg = `linear-gradient(90deg, var(--bg-value-color) 0%, var(--bg-value-color) ${percent100}%, var(--bg-range-color) ${percent100}%, var(--bg-range-color) 100%)`;
-            this.$el.querySelector('input').style.backgroundImage = bg;
-        }
-    }
-}
+  },
+};
 </script>
-  
+
 <style scoped>
 .bpm-slider {
     /* height: 0px;
@@ -132,4 +132,4 @@ input {
     opacity: 1;
 }
 </style>
-  
+
