@@ -1,15 +1,15 @@
-import Vue from "vue";
-import store from "./store/index.js";
-import App from "./App.vue";
-import router from "./router";
-import Vue2TouchEvents from "vue2-touch-events";
-import ToggleButton from 'vue-js-toggle-button'
-import VModal from 'vue-js-modal'
+import Vue from 'vue';
+import store from './store/index.js';
+import App from './App.vue';
+import router from './router';
+import Vue2TouchEvents from 'vue2-touch-events';
+import ToggleButton from 'vue-js-toggle-button';
+import VModal from 'vue-js-modal';
 import Toasted from 'vue-toasted';
-import { MdButton, MdIcon, MdField} from 'vue-material/dist/components'
-import 'vue-material/dist/vue-material.min.css'
-import 'vue-material/dist/theme/default.css'
-import "remixicon/fonts/remixicon.css";
+import {MdButton, MdIcon, MdField} from 'vue-material/dist/components';
+import 'vue-material/dist/vue-material.min.css';
+import 'vue-material/dist/theme/default.css';
+import 'remixicon/fonts/remixicon.css';
 import Dropdown from 'vue-simple-search-dropdown';
 
 Vue.use(MdButton);
@@ -18,17 +18,17 @@ Vue.use(MdField);
 Vue.use(Dropdown);
 Vue.use(VModal);
 Vue.use(Toasted, {
-  duration: 1000
+  duration: 1000,
 });
 Vue.use(ToggleButton);
 Vue.use(Vue2TouchEvents, {
   disableClick: true,
-  touchClass: "",
+  touchClass: '',
   tapTolerance: 10,
   touchHoldTolerance: 400,
   swipeTolerance: 30,
   longTapTimeInterval: 400,
-  namespace: "touch",
+  namespace: 'touch',
 });
 Vue.config.productionTip = false;
 
@@ -36,4 +36,4 @@ new Vue({
   router,
   store,
   render: (h) => h(App),
-}).$mount("#app");
+}).$mount('#app');
