@@ -1,10 +1,10 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
+import Vue from 'vue';
+import Vuex from 'vuex';
 // import pianobufferstate from './modules/piano-buffer-state'
-import noteBuffers from './modules/note-buffers'
-import tickNumber from './modules/tick-number'
-import globalSettings from './modules/global-settings'
-import samplers from './modules/samplers'
+import noteBuffers from './modules/note-buffers';
+import tickNumber from './modules/tick-number';
+import globalSettings from './modules/global-settings';
+import samplers from './modules/samplers';
 // import newmodule from './modules/newmodule'
 
 Vue.use(Vuex);
@@ -26,8 +26,8 @@ export default new Vuex.Store({
     noteBuffers,
     tickNumber,
     globalSettings,
-    samplers
+    samplers,
     // add new modules here in this manner:
     // newmodule,
   },
-})
+});
