@@ -1,112 +1,112 @@
 class Node {
-    constructor(value) {
-      this.value = value;
-      this.next = null;
+  constructor(value) {
+    this.value = value;
+    this.next = null;
+  }
+}
+
+class LIFOQueue {
+  constructor(maxLength) {
+    this.head = null;
+    this.maxLength = maxLength;
+    this.size = 0;
+  }
+
+  push(element) {
+    const newNode = new Node(element);
+    newNode.next = this.head;
+    this.head = newNode;
+    this.size++;
+
+    if (this.size > this.maxLength) {
+      this.pop(); // Remove the oldest element
     }
   }
-  
-class LIFOQueue {
-    constructor(maxLength) {
-        this.head = null;
-        this.maxLength = maxLength;
-        this.size = 0;
-    }
 
-    push(element) {
-        const newNode = new Node(element);
-        newNode.next = this.head;
-        this.head = newNode;
-        this.size++;
-
-        if (this.size > this.maxLength) {
-            this.pop(); // Remove the oldest element
-        }
+  pop() {
+    if (this.head === null) {
+      return undefined; // Queue is empty
     }
+    const poppedValue = this.head.value;
+    this.head = this.head.next;
+    this.size--;
 
-    pop() {
-        if (this.head === null) {
-            return undefined; // Queue is empty
-        }
-        const poppedValue = this.head.value;
-        this.head = this.head.next;
-        this.size--;
+    return poppedValue;
+  }
 
-        return poppedValue;
-    }
-    
-    length() {
-        return this.size;
-    }
+  length() {
+    return this.size;
+  }
 
-    clear() {
-        this.head = null;
-        this.size = 0;
-      }
-    
-    toArray() {
-        const arr = [];
-        let current = this.head;
-        while (current !== null) {
-            arr.push(current.value);
-            current = current.next;
-        }
-        return arr;
+  clear() {
+    this.head = null;
+    this.size = 0;
+  }
+
+  toArray() {
+    const arr = [];
+    let current = this.head;
+    while (current !== null) {
+      arr.push(current.value);
+      current = current.next;
     }
-    toArrayAndClear() {
-        const arr = [];
-        while (this.head !== null) {
-          arr.push(this.pop());
-        }
-        return arr;
-      }
+    return arr;
+  }
+  toArrayAndClear() {
+    const arr = [];
+    while (this.head !== null) {
+      arr.push(this.pop());
+    }
+    return arr;
+  }
 }
 
 class FIFOQueue {
-    constructor(maxLength) {
-        this.head = null;
-        this.tail = null;
-        this.maxLength = maxLength;
-        this.size = 0;
+  constructor(maxLength) {
+    this.head = null;
+    this.tail = null;
+    this.maxLength = maxLength;
+    this.size = 0;
+  }
+
+  push(element) {
+    const newNode = new Node(element);
+
+    if (this.head === null) {
+      this.head = newNode;
+      this.tail = newNode;
+    } else {
+      this.tail.next = newNode;
+      this.tail = newNode;
     }
 
-    push(element) {
-        const newNode = new Node(element);
+    this.size++;
 
-        if (this.head === null) {
-            this.head = newNode;
-            this.tail = newNode;
-        } else {
-            this.tail.next = newNode;
-            this.tail = newNode;
-        }
+    if (this.size > this.maxLength) {
+      this.pop(); // Remove the oldest element
+    }
+  }
 
-        this.size++;
-
-        if (this.size > this.maxLength) {
-            this.pop(); // Remove the oldest element
-        }
+  pop() {
+    if (this.head === null) {
+      return undefined; // Queue is empty
     }
 
-    pop() {
-        if (this.head === null) {
-            return undefined; // Queue is empty
-        }
+    const poppedValue = this.head.value;
+    this.head = this.head.next;
 
-        const poppedValue = this.head.value;
-        this.head = this.head.next;
-
-        if (this.head === null) {
-            this.tail = null;
-        }
-
-        this.size--;
-
-        return poppedValue;
+    if (this.head === null) {
+      this.tail = null;
     }
 
-    length() {
-        return this.size;
-    }
+    this.size--;
+
+    return poppedValue;
+  }
+
+  length() {
+    return this.size;
+  }
 }
 
 export {LIFOQueue, FIFOQueue};
