@@ -1,24 +1,24 @@
-import { note, transpose } from "@tonaljs/tonal"
-import { enharmonic } from "@tonaljs/note"
+import {note, transpose} from '@tonaljs/tonal';
+import {enharmonic} from '@tonaljs/note';
 
 export function createRange(from, to) {
-  let fromNote = note(from)
-  const toNote = note(to)
+  let fromNote = note(from);
+  const toNote = note(to);
 
   if (fromNote.height >= toNote.height) {
-    throw new Error("Reverse ranges are not yet implemented.")
+    throw new Error('Reverse ranges are not yet implemented.');
   }
 
-  if (fromNote.acc === "b") {
-    fromNote = note(enharmonic(fromNote))
+  if (fromNote.acc === 'b') {
+    fromNote = note(enharmonic(fromNote));
   }
 
-  let range = []
+  const range = [];
 
   for (let i = 0, l = toNote.height - fromNote.height, currNote = fromNote; i < l; i++) {
-    range.push(currNote)
-    currNote = note(enharmonic(transpose(currNote, "m2")))
+    range.push(currNote);
+    currNote = note(enharmonic(transpose(currNote, 'm2')));
   }
 
-  return range
+  return range;
 }
