@@ -2,41 +2,41 @@
 /*
 Import custom components
 */
-import Keyboard from "@/components/Keyboard.vue";
-import PianoRoll from "@/components/PianoRollLegacy.vue";
-import Score from "@/components/Score.vue";
-import VerticalSlider from '@/components/VerticalSlider.vue'
+import Keyboard from '@/components/Keyboard.vue';
+import PianoRoll from '@/components/PianoRollLegacy.vue';
+import Score from '@/components/Score.vue';
+import VerticalSlider from '@/components/VerticalSlider.vue';
 // import HorizontalSlider from '@/components/HorizontalSlider.vue'
-import AudioMeter from "@/components/AudioMeter.vue";
-import BPMSlider from '@/components/BPMSlider.vue'
+import AudioMeter from '@/components/AudioMeter.vue';
+import BPMSlider from '@/components/BPMSlider.vue';
 // import VectorBar from "../components/VectorBar.vue";
-import ChromaChart from "@/components/ChromaChart.vue";
-import Monitor from "@/components/Monitor.vue";
-import Mixer from "@/components/Mixer.vue";
-import TextBox from "@/components/TextBox.vue";
+import ChromaChart from '@/components/ChromaChart.vue';
+import Monitor from '@/components/Monitor.vue';
+import Mixer from '@/components/Mixer.vue';
+import TextBox from '@/components/TextBox.vue';
 
-import "../css/main.css";
-import * as Tone from "tone";
-import { Midi, note } from "@tonaljs/tonal";
-import { WebMidi } from "webmidi";
-import Dropdown from "vue-simple-search-dropdown";
-import AudioKeys from "audiokeys";
+import '../css/main.css';
+import * as Tone from 'tone';
+import {Midi} from '@tonaljs/tonal';
+import {WebMidi} from 'webmidi';
+import Dropdown from 'vue-simple-search-dropdown';
+import AudioKeys from 'audiokeys';
 
-import yaml from "js-yaml";
-import * as rb from "ringbuf.js";
+import yaml from 'js-yaml';
+import * as rb from 'ringbuf.js';
 import {
     playerType, instrumentType, eventSourceType,
     messageType, statusType, noteType,
     uiParameterType,
-    agentHookType
-} from '@/utils/types.js'
+    agentHookType,
+} from '@/utils/types.js';
 
-import { URLFromFiles, isMobile, isNotChrome } from '@/utils/helpers.js'
-import { NoteEvent } from '@/utils/NoteEvent.js'
+import {urlFromFiles, isMobile, isNotChrome} from '@/utils/helpers.js';
+import {NoteEvent} from '@/utils/NoteEvent.js';
 
 export default {
 
-    name: "mainScreen",
+    name: 'mainScreen',
 
     components: {
         Keyboard,
@@ -49,24 +49,23 @@ export default {
         TextBox,
         Mixer,
         Monitor,
-        ChromaChart
+        ChromaChart,
         // VectorBar,
         // HorizontalSlider,
-
     },
 
     data() {
         return {
-            // Choose the agent. 
+            // Choose the agent.
             // This string should be one of
             // dir names inside public/agents/
-            agentName: "meydaDemo",
+            agentName: 'meydaDemo',
             // Provide all the config files that should be loaded
             // These should be in public/agents/{agentName}/
-            configFiles: ['config.yaml', 'config_widgets.yaml', 'config_players.yaml'],
-
+            configFiles: ['config.yaml',
+                'config_widgets.yaml',
+                'config_players.yaml'],
             config: null,
-
             playerType,
             instrumentType,
             eventSourceType,
@@ -75,13 +74,13 @@ export default {
             noteType,
             uiParameterType,
             agentHookType,
-
-            // These need to be initialized here as empty arrays (computed properties)
+            // These need to be initialized here
+            // as empty arrays (computed properties)
             // They'll be filled in the created() hook
             switches: [],
             sliders: [],
             buttons: [],
-            // Information about the players and 
+            // Information about the players and
             // instruments available to each player.
             // It's used to create the mixer modal.
             // This is filled in the created() hook
@@ -91,7 +90,8 @@ export default {
 
             // Score status
             scoreShown: true,
-            scrollStatus: true, // THIS ONE. Too many flags for score
+            // TODO: Too many flags for score
+            scrollStatus: true,
             scoreStatus: true,
             // Textbox status
             textBoxTitle: null,
@@ -99,7 +99,8 @@ export default {
             textBoxStatus: false,
 
             localBPM: null,
-            localSyncClockStatus: false, // used to trigger local UI change
+            // used to trigger local UI change
+            localSyncClockStatus: false,
             screenWidth: document.body.clientWidth,
             screenHeight: document.body.clientHeight,
             keyboardKey: 0,
@@ -127,22 +128,25 @@ export default {
             pageLoadTime: null,
             modelLoadTime: null,
             activeDevices: [],
-            selectedMIDIDevice: "",
+            selectedMIDIDevice: '',
 
             noteOffEventForNextTick: null,
 
-            // used to calculate the average agent inference time (gridBased mode) 
+            // used to calculate the average
+            // agent inference time (gridBased mode)
             // and estimate maxBPM
             modelInferenceTimes: [],
             // maxBPM (or min clock period) supported by the current device
             maxBPM: 0,
-            // counter for the number of times the agent inference time exceeds the clock period
+            // counter for the number of times the agent
+            // inference time exceeds the clock period
             misalignErrCount: 0,
 
             isNotChrome,
             isMobile,
-            // Keep track of all the timeouts ids to clear them when the the user pauses
-            timeout_IDS_kill: [],// noteOn related events (keyDown, mouseDown, noteOn, trigerAttack etc)
+            // Keep track of all the timeouts ids to
+            // clear them when the the user pauses
+            timeout_IDS_kill: [],
             timeout_IDS_live: [],
 
             mixer_data: null,
@@ -150,24 +154,25 @@ export default {
     },
 
     created() {
-        let vm = this;
+        const vm = this;
 
-        console.log("created main start")
+        console.log('created main start');
         this.loadConfigSync();
-        console.log("load config sync done")
+        console.log('load config sync done');
 
-        this.$store.commit("setConfig", this.config);
-        this.$store.commit("initQuantBuffers", this.config);
-        this.$store.commit("setTicksPerMeasure", this.config);
-        this.$store.commit("createInstruments", this.config);
-        this.$store.commit("setNoteType", this.noteType);
+        this.$store.commit('setConfig', this.config);
+        this.$store.commit('initQuantBuffers', this.config);
+        this.$store.commit('setTicksPerMeasure', this.config);
+        this.$store.commit('createInstruments', this.config);
+        this.$store.commit('setNoteType', this.noteType);
 
         // Activate/Deactivate GUI widgets based on config
-        this.scoreStatus = this.config.gui.score.status
-        this.textBoxStatus = this.config.gui.textBox.status
+        this.scoreStatus = this.config.gui.score.status;
+        this.textBoxStatus = this.config.gui.textBox.status;
 
         // Update the score properties
-        this.scoreShown = this.config.gui.score.status; //TODO: unclear use in Score.vue
+        // TODO: unclear use in Score.vue
+        this.scoreShown = this.config.gui.score.status;
         // this.scrollEnabled = this.config.gui.score;
         // Set the textBox title
         this.textBoxTitle = this.config.gui.textBox.title;
@@ -185,53 +190,55 @@ export default {
         }
 
         this.players = this.config.players;
-        
+
         // if (vm.config.gui.monitor && this.config.gui.monitor.status){
-        if (this.showMonitor){
+        if (this.showMonitor) {
             this.monitorObserverInterval = 10;
-            if (this.config.gui.monitor.structure)
+            if (this.config.gui.monitor.structure) {
                 this.config.gui.monitor.structure.forEach((tab) => {
                     tab.parameters.forEach((parameter) => {
                         vm.dataForMonitoring[parameter.id] = 0;
                     });
                 });
+            }
         }
-        console.log("created localBPM set to defaultBPM from config ", this.config.clockSettings.defaultBPM);
         this.localBPM = this.config.clockSettings.defaultBPM;
-        console.log("created main end")
+        console.log('created main end');
     },
 
     beforeMount() {
-        console.log("beforeMount main start");
+        console.log('beforeMount main start');
         // let vm = this;
-        console.log("beforeMount main end");
+        console.log('beforeMount main end');
     },
 
     async mounted() {
-        console.log("main mounted");
-        var vm = this;
-        if (vm.showMonitor){
-            console.log("MA DEN EPREPE")
+        console.log('main mounted');
+        const vm = this;
+        if (vm.showMonitor) {
+            console.log('MA DEN EPREPE');
             vm.$root.$refs.monitor.loadMonitorConfig(vm.config.gui.monitor);
         }
         vm.$root.$refs.mixer.loadMixerConfig(vm.config.players);
 
         /*
-        * Loading Animation: set initial status of both div
-        */
-        vm.$refs.mainContent.style.display = "none";
-        vm.$refs.entryBtn.style.visibility = "hidden";
+            * Loading Animation: set initial status of both div
+            */
+        vm.$refs.mainContent.style.display = 'none';
+        vm.$refs.entryBtn.style.visibility = 'hidden';
 
         vm.audioContext = new AudioContext();
-        // Tone.setContext(vm.audioContext); // this cause huge latency
-        Tone.context.lookAhead = 0; // increase it if you experience audio clicks
+        // this cause huge latency for some reason
+        // Tone.setContext(vm.audioContext);
+        // increase it if you experience audio clicks
+        Tone.context.lookAhead = 0;
 
         // SAB
         // get a memory region for the Audio ring buffer
         // length in time is 1 second of stereo audio
         // Float32Array is 4 bytes per sample
-        vm.sab = rb.RingBuffer.getStorageForCapacity(vm.audioContext.sampleRate * 2, Float32Array);
-
+        vm.sab = rb.RingBuffer.getStorageForCapacity(
+            vm.audioContext.sampleRate * 2, Float32Array);
 
         // get a memory region for the parameter ring buffer
         // This one is to send parameters from the UI to the agent
@@ -247,18 +254,11 @@ export default {
 
 
         // Initialize agent worker
-        // experiment with , { type : 'module' }
-        // if agen'ts name is pianoGenie, then load the pianoGenie worker
-        // if (vm.agentName == "pianoGenie") {
-        vm.agent = new Worker(`/agents/${vm.agentName}/agent.js`, { type: 'module' });
-        // } else {
-        //     vm.agent = new Worker(`/agents/${vm.agentName}/agent.js`);
-        // }
-        // vm.agent = new Worker(`/agents/${vm.agentName}/agent.js`,  { type : 'module' });
+        vm.agent = new Worker(`/agents/${vm.agentName}/agent.js`, {type: 'module'});
 
         vm.agent.onmessage = vm.agentCallback;
 
-        // Send a message to agent with some necessary 
+        // Send a message to agent with some necessary
         // configurations and constants to store inside the agent
         // TODO : i removed await
         vm.agent.postMessage({
@@ -277,22 +277,23 @@ export default {
                 sab_par_ui: vm.sab_par_ui,
                 sab_par_agent: vm.sab_par_agent,
                 channelCount: 2,
-                sampleRate: vm.audioContext.sampleRate
-            }
+                sampleRate: vm.audioContext.sampleRate,
+            },
         });
 
-        if (vm.showMonitor){
-            if (vm.config.gui.monitor.structure){
-                vm.timeout_IDS_live.push(setInterval(vm.agentParameterObserver, vm.monitorObserverInterval));
+        if (vm.showMonitor) {
+            if (vm.config.gui.monitor.structure) {
+                vm.timeout_IDS_live.push(
+                    setInterval(vm.agentParameterObserver, vm.monitorObserverInterval));
             }
         }
 
 
         // Initialize Clock Worker (module)
-        vm.clockWorker = new Worker("/clock.js", { type: "module" });
+        vm.clockWorker = new Worker('/clock.js', {type: 'module'});
         vm.clockWorker.onmessage = vm.tickBehavior;
-        vm.clockWorker.postMessage({ action: 'setBPM', bpm: vm.localBPM });
-        vm.$store.commit("initializeClock");
+        vm.clockWorker.postMessage({action: 'setBPM', bpm: vm.localBPM});
+        vm.$store.commit('initializeClock');
 
         if (vm.config.interactionMode.audioMode) {
             /*
@@ -304,7 +305,7 @@ export default {
             });
 
             vm.mediaStreamSource = vm.audioContext.createMediaStreamSource(
-                stream
+                stream,
             );
 
 
@@ -321,21 +322,22 @@ export default {
             // vm.$root.$refs.vectorBar.updateAnalysis();
             vm.audioContext.resume(); // ?
 
-            const recorderWorkletUrl = await URLFromFiles(['recorder-worklet.js', 'libraries/index_rb.js'])
+            const recorderWorkletUrl = await urlFromFiles(
+                ['recorder-worklet.js', 'libraries/index_rb.js'],
+            );
             await vm.audioContext.audioWorklet.addModule(recorderWorkletUrl);
-
 
 
             vm.recorderWorkletNode = new AudioWorkletNode(
                 vm.audioContext,
-                "recorder-worklet",
-                { processorOptions: vm.sab }
+                'recorder-worklet',
+                {processorOptions: vm.sab},
             );
 
-            vm.recorderWorkletNode.port.postMessage("ping")
+            vm.recorderWorkletNode.port.postMessage('ping');
 
-            vm.recorderWorkletNode.port.addEventListener("message", (event) => {
-                console.log("Received from Worklet" + event.data);
+            vm.recorderWorkletNode.port.addEventListener('message', (event) => {
+                console.log('Received from Worklet' + event.data);
             });
             // vm.recorderWorkletNode.port.start(); # TODO do I need this for ping/pong ?
             // send the mic to the recorderNode --> recorderWorklet
@@ -347,32 +349,32 @@ export default {
         * Web MIDI logic
         */
         if (navigator.requestMIDIAccess) {
-            navigator.requestMIDIAccess().then(function (access) {
+            navigator.requestMIDIAccess().then(function(access) {
                 vm.WebMIDISupport = true;
                 access.onstatechange = vm.onEnabled;
             });
             // Enable WebMIDI, then call onEnabled method.
             WebMidi.enable()
                 .then(vm.onEnabled)
-                .catch((err) => this.$toasted.show("WebMIDI Error: " + err));
+                .catch((err) => this.$toasted.show('WebMIDI Error: ' + err));
         }
 
         /*
         * Initialize computer keyboard logic
         * AudioKeys maps the computer keyboard to midi values
         */
-        var keyboard = new AudioKeys({
-            // set polyphony it to a very high number. 
+        const keyboard = new AudioKeys({
+            // set polyphony it to a very high number.
             // we will handle the polyphony ourselves.
             polyphony: 100,
             rows: 2,
-            priority: "last",
+            priority: 'last',
             rootNote: 60,
         });
 
         // callback for when a laptop keyboard key is pressed
-        keyboard.down(function (note) {
-            let noteName = Midi.midiToNoteName(note.note, { sharps: true });
+        keyboard.down(function(note) {
+            const noteName = Midi.midiToNoteName(note.note, {sharps: true});
             const newNoteEvent = new NoteEvent();
             newNoteEvent.player = vm.playerType.HUMAN;
             newNoteEvent.instrument = vm.instrumentType.PIANO;
@@ -388,15 +390,15 @@ export default {
             };
             newNoteEvent.playAfter = {
                 seconds: 0,
-                tick: 0
+                tick: 0,
             };
             newNoteEvent.duration = null;
             vm.processUserNoteEvent(newNoteEvent);
         });
 
         // callback for when a laptop keyboard key is released
-        keyboard.up(function (note) {
-            let noteName = Midi.midiToNoteName(note.note, { sharps: true });
+        keyboard.up(function(note) {
+            const noteName = Midi.midiToNoteName(note.note, {sharps: true});
             const newNoteEvent = new NoteEvent();
             newNoteEvent.player = vm.playerType.HUMAN;
             newNoteEvent.instrument = vm.instrumentType.PIANO;
@@ -412,11 +414,11 @@ export default {
             };
             newNoteEvent.playAfter = {
                 seconds: 0,
-                tick: 0
+                tick: 0,
             };
             newNoteEvent.duration = null;
 
-            vm.processUserNoteEvent(newNoteEvent)
+            vm.processUserNoteEvent(newNoteEvent);
         });
 
         /*
@@ -428,36 +430,37 @@ export default {
                 vm.screenWidth = window.screenWidth;
             })();
         };
+
         // Prevent spacebar trigger any button
-        document.querySelectorAll("button").forEach(function (item) {
-            item.addEventListener("focus", function () {
+        document.querySelectorAll('button').forEach(function(item) {
+            item.addEventListener('focus', function() {
                 this.blur();
             });
         });
 
         // spacebar trigger play btn
-        document.addEventListener("keypress", function (event) {
-            if (event.code == "Space" && !vm.$store.getters.getModalStatus) {
+        document.addEventListener('keypress', function(event) {
+            if (event.code == 'Space' && !vm.$store.getters.getModalStatus) {
                 // spacebar could toggle clock
                 vm.toggleClock();
             }
         });
 
         vm.modelLoadTime = Date.now();
-        console.log("TONE TONE TONE ", Tone.now());
-
+        console.log('TONE TONE TONE ', Tone.now());
     },
     methods: {
 
         // Clock behavior function.
         async tickBehavior() {
-            var vm = this;
+            const vm = this;
             if (vm.$store.getters.getClockStatus) {
-                vm.$store.commit("incrementTick");
+                vm.$store.commit('incrementTick');
 
                 // Trigger the metronome only if there is a "metronome" entry in config_players.yaml
-                if (vm.config.players.metronome)
+                if (vm.config.players.metronome) {
                     vm.metronomeTrigger();
+                }
 
                 vm.calculateMaxBPM();
 
@@ -465,18 +468,19 @@ export default {
                 vm.triggerAgentSamplerSync(); // UNCOMMENT
 
 
-                // run the agent with a small delay of tick/4 in order to include 
+                // run the agent with a small delay of tick/4 in order to include
                 // any notes that the user played very close to the tick change.
                 // this makes the grid a bit more flexible, and the human input is correctly parsed
-                // In terms of playability, the human finds it much more easy to play along the metronome on the grid
+                // In terms of playability, the human finds it much more easy to play
+                // along the metronome on the grid
                 // TUTOR: have a flag for that in config named delayedExecution
                 if (vm.config.noteModeSettings.gridBased.status) {
                     if (vm.config.noteModeSettings.gridBased.delayedExecution) {
                         // console.log("delayedExecution");
-                        vm.timeout_IDS_live.push(setTimeout(function () {
+                        vm.timeout_IDS_live.push(setTimeout(function() {
                             vm.estimateHumanQuantizedNote();
                             vm.runTheAgent();
-                        }, parseInt(vm.$store.getters.getClockPeriod / 4))
+                        }, parseInt(vm.$store.getters.getClockPeriod / 4)),
                         );
                     } else {
                         vm.estimateHumanQuantizedNote();
@@ -484,102 +488,107 @@ export default {
                     }
                 } else {
                     // inside runTheAgent we increment the "delayed" tick number.
-                    // If we don't run the agent, we still want to increment the "delayed" tick number
-                    // because other parts of the code depend on it. In this case tick === delayedTick
-                    vm.$store.commit("incrementTickDelayed");
+                    // If we don't run the agent, we still want to increment the
+                    // "delayed" tick number because other parts of the code
+                    // depend on it. In this case tick === delayedTick
+                    vm.$store.commit('incrementTickDelayed');
                 }
             }
         },
 
         async toggleClock() {
-            var vm = this;
+            const vm = this;
             if (!vm.localSyncClockStatus) {
                 vm.startRecording();
-                vm.clockWorker.postMessage({ action: 'startClock' });
+                vm.clockWorker.postMessage({action: 'startClock'});
             } else {
                 vm.stopRecording();
-                vm.clockWorker.postMessage({ action: 'stopClock' });
+                vm.clockWorker.postMessage({action: 'stopClock'});
             }
             vm.localSyncClockStatus = !vm.localSyncClockStatus;
-            vm.$store.commit("changeClockStatus");
-            vm.$store.commit("clearPianoState");
+            vm.$store.commit('changeClockStatus');
+            vm.$store.commit('clearPianoState');
             // vm.manuallyUpdateData();
         },
 
         stopClock() {
             this.stopRecording();
-            this.clockWorker.postMessage({ action: 'stopClock' });
+            this.clockWorker.postMessage({action: 'stopClock'});
             this.localSyncClockStatus = false;
-            this.$store.commit("setClockStatus", false);
-            this.$store.commit("clearPianoState");
+            this.$store.commit('setClockStatus', false);
+            this.$store.commit('clearPianoState');
         },
 
         startClock() {
             this.startRecording();
-            this.clockWorker.postMessage({ action: 'startClock' });
+            this.clockWorker.postMessage({action: 'startClock'});
             this.localSyncClockStatus = true;
-            this.$store.commit("setClockStatus", true);
-            this.$store.commit("clearPianoState");
+            this.$store.commit('setClockStatus', true);
+            this.$store.commit('clearPianoState');
         },
 
         metronomeTrigger() {
-            var vm = this;
+            const vm = this;
             // This method would trigger the metronome sampler.
             // getTicksPerBeat returns the number of ticks per beat
             // and we want the metronome to trigger every beat
             if (this.$store.getters.getLocalTick % this.$store.getters.getTicksPerBeat == 0) {
-                var currentNote =
-                    this.$store.getters.getLocalTick % this.$store.getters.getTicksPerMeasure === 0 ? "G0" : "C0";
-                const metronomeNote = new NoteEvent()
+                const currentNote = this.$store.getters.getLocalTick %
+                        this.$store.getters.getTicksPerMeasure === 0 ? 'G0' : 'C0';
+                const metronomeNote = new NoteEvent();
                 metronomeNote.player = vm.playerType.METRONOME;
                 metronomeNote.name = currentNote;
                 metronomeNote.type = vm.noteType.NOTE_ON;
                 metronomeNote.velocity = 127;
                 metronomeNote.playAfter = {
                     tick: 0,
-                    seconds: 0
-                }
-                this.$store.dispatch("samplerOn", metronomeNote);
-
+                    seconds: 0,
+                };
+                this.$store.dispatch('samplerOn', metronomeNote);
             }
         },
 
         processUserNoteEvent(noteEvent, onScreenKeyboard = false) {
             // return 0;
-            var vm = this;
+            const vm = this;
             // We always send the user's input directly to the sampler
             // for immediate playback
-            let whiteKey = noteEvent.name.includes('#') ? false : true;
+            const whiteKey = noteEvent.name.includes('#') ? false : true;
             // Check if the current on-screen key is within the current screen view
-            // if (this.$root.$refs.keyboard.$refs[noteEvent.name] ) is null then the key is not on screen
-            let keyOnScreenRange = this.$root.$refs.keyboard.$refs[noteEvent.name] ? true : false;
+            // if (this.$root.$refs.keyboard.$refs[noteEvent.name] )
+            // is null then the key is not on screen
+            const keyOnScreenRange = this.$root.$refs.keyboard.$refs[noteEvent.name] ? true : false;
             if (noteEvent.type == vm.noteType.NOTE_ON) {
                 // console.log("note on");
                 if (this.config.gui.pianoRoll.status && this.config.gui.pianoRoll.human) {
                     vm.$root.$refs.pianoRoll.keyDown(noteEvent);
                 }
-                vm.$store.dispatch("noteOn", noteEvent);
-                vm.$store.dispatch("samplerOn", noteEvent);
+                vm.$store.dispatch('noteOn', noteEvent);
+                vm.$store.dispatch('samplerOn', noteEvent);
 
                 if (keyOnScreenRange && !onScreenKeyboard) {
                     if (whiteKey) {
-                        this.$root.$refs.keyboard.$refs[noteEvent.name][0].classList.add('active-white-key-human')
+                        this.$root.$refs.keyboard.$refs[noteEvent.name][0]
+                            .classList.add('active-white-key-human');
                     } else {
-                        this.$root.$refs.keyboard.$refs[noteEvent.name][0].classList.add('active-black-key-human')
+                        this.$root.$refs.keyboard.$refs[noteEvent.name][0]
+                            .classList.add('active-black-key-human');
                     }
                 }
             } else if (noteEvent.type == vm.noteType.NOTE_OFF) {
                 if (this.config.gui.pianoRoll.status && this.config.gui.pianoRoll.human) {
                     vm.$root.$refs.pianoRoll.keyUp(noteEvent);
                 }
-                vm.$store.dispatch("noteOff", noteEvent);
-                vm.$store.dispatch("samplerOff", noteEvent);
+                vm.$store.dispatch('noteOff', noteEvent);
+                vm.$store.dispatch('samplerOff', noteEvent);
 
                 if (keyOnScreenRange && !onScreenKeyboard) {
                     if (whiteKey) {
-                        this.$root.$refs.keyboard.$refs[noteEvent.name][0].classList.remove('active-white-key-human')
+                        this.$root.$refs.keyboard.$refs[noteEvent.name][0]
+                            .classList.remove('active-white-key-human');
                     } else {
-                        this.$root.$refs.keyboard.$refs[noteEvent.name][0].classList.remove('active-black-key-human')
+                        this.$root.$refs.keyboard.$refs[noteEvent.name][0]
+                            .classList.remove('active-black-key-human');
                     }
                 }
             }
@@ -604,38 +613,43 @@ export default {
             // console.log("runningTheagent");
             // remember, runTheAgent might happen with a small delay of tick/4 after the tick
             // here I just keep track of the 'delayed' tick
-            this.$store.commit("incrementTickDelayed");
+            this.$store.commit('incrementTickDelayed');
 
             // MAJOR TODO : draw should probably go before the delayedTickIncrement
             if (vm.config.gui.score.status) {
                 this.$root.$refs.score.draw();
-                let lastNote = this.$store.getters.getLastHumanNoteQuantized
-                // console.log("startTick", lastNote.startTick, "midi", lastNote.midi, "duration", lastNote.dur);
             }
-            let messageContent = {
+            const messageContent = {
                 tick: this.$store.getters.getLocalTick,
                 globalTick: this.$store.getters.getGlobalTick,
+            };
+            if (vm.config.noteModeSettings.gridBased.eventBuffer) {
+                messageContent['humanContinuousBuffer'] = this.$store.getters
+                    .getMidiEventBuffer;
             }
-            if (vm.config.noteModeSettings.gridBased.eventBuffer)
-                messageContent['humanContinuousBuffer'] = this.$store.getters.getMidiEventBuffer;
-            if (vm.config.noteModeSettings.gridBased.quantizedEvents)
-                messageContent['humanQuantizedInput'] = this.$store.getters.getHumanInputFor(this.$store.getters.getLocalTick);
+            if (vm.config.noteModeSettings.gridBased.quantizedEvents) {
+                messageContent['humanQuantizedInput'] = this.$store.getters
+                    .getHumanInputFor(this.$store.getters.getLocalTick);
+            }
             this.agent.postMessage({
                 hookType: vm.agentHookType.CLOCK_EVENT,
                 content: messageContent,
-            })
+            });
         },
 
         estimateHumanQuantizedNote() {
-            var vm = this;
-            /* 
-            1) Get the raw continuous buffers for the current tick (bufferOn, bufferOff, bufferEvent)
-            2) preprocess them before quantization
-                a) In bufferEvent, we can remove the notes whose duration is less than the tick duration 
+            const vm = this;
+            /*
+                1) Get the raw continuous buffers for the current tick
+                    (bufferOn, bufferOff, bufferEvent)
+                2) preprocess them before quantization
+                    a) In bufferEvent, we can remove the notes whose duration
+                    is less than the tick duration
             */
 
             const bufferEvent = this.$store.getters.getMidiEventBuffer;
-            // A this is the noteOff for a note from the previous tick that was too short to be quantized
+            // A this is the noteOff for a note from the previous
+            // tick that was too short to be quantized
             if (vm.noteOffEventForNextTick) {
                 bufferEvent.push(vm.noteOffEventForNextTick);
             }
@@ -645,18 +659,21 @@ export default {
             //     console.log("bufferEvent", bufferEvent);
             // }
             // a COPY activePianoNotes are sorted by their "on" timestamp (newest to oldest)
-            let activePianoNotes = [...this.$store.getters.getActivePianoNotes];
-            let currentQuantizedEvents = [];
+            const activePianoNotes = [...this.$store.getters.getActivePianoNotes];
+            const currentQuantizedEvents = [];
 
             // 2) preprocess them before quantization
-            //     a) In bufferEvent, we can remove the notes whose duration is less than the tick duration 
+            //     a) In bufferEvent, we can remove the notes whose
+            //        duration is less than the tick duration
             // TODO : move this to utilities or smth
             let indexesToRemove = []; // the indexes of the events to be removed
             // for (let i = bufferEvent.length - 1; i >= 0; i--) {
             //     let elem = bufferEvent[i];
             //     if (elem.type === vm.noteType.NOTE_OFF) {
             //         const matchingIndexes = bufferEvent
-            //             .map((e, i) => (e.type === vm.noteType.NOTE_ON && e.midi === elem.midi && e.createdAt.seconds < elem.createdAt.seconds) ? i : -1)
+            //             .map((e, i) => (e.type === vm.noteType.NOTE_ON &&
+            //                              e.midi === elem.midi &&
+            //                              e.createdAt.seconds < elem.createdAt.seconds) ? i : -1)
             //             .filter(index => index !== -1);
             //         if (matchingIndexes.length > 0) {
             //             indexesToRemove.push(i);
@@ -667,11 +684,13 @@ export default {
             let forgivenNoteOnIndex = -1;
             vm.noteOffEventForNextTick = null;
             for (let i = bufferEvent.length - 1; i >= 0; i--) {
-                let elem = bufferEvent[i];
+                const elem = bufferEvent[i];
                 if (elem.type === vm.noteType.NOTE_OFF) {
                     const matchingIndexes = bufferEvent
-                        .map((e, i) => (e.type === vm.noteType.NOTE_ON && e.midi === elem.midi && e.createdAt.seconds < elem.createdAt.seconds) ? i : -1)
-                        .filter(index => index !== -1);
+                        .map((e, i) => (e.type === vm.noteType.NOTE_ON &&
+                            e.midi === elem.midi &&
+                            e.createdAt.seconds < elem.createdAt.seconds) ? i : -1)
+                        .filter((index) => index !== -1);
                     if (matchingIndexes.length > 0) {
                         indexesToRemove.push(i);
                         forgivenNoteOnIndex = Math.max(...matchingIndexes);
@@ -683,40 +702,38 @@ export default {
             indexesToRemove = Array.from(new Set([...indexesToRemove]));
             if (indexesToRemove.length > 0) {
                 if (indexesToRemove.length > 1) {
-                    console.warn("indexesToRemove", indexesToRemove);
+                    console.warn('indexesToRemove', indexesToRemove);
                     // TODO : this can happen when playing a whole chord with duration
                     // less than the tick.
                 }
                 if (forgivenNoteOnIndex == -1) {
-                    console.error("forgivenNoteOnIndex", forgivenNoteOnIndex);
+                    console.error('forgivenNoteOnIndex', forgivenNoteOnIndex);
                 }
                 vm.noteOffEventForNextTick = bufferEvent[indexesToRemove[0]];
-                let forgivenNoteOn = bufferEvent[forgivenNoteOnIndex];
+                const forgivenNoteOn = bufferEvent[forgivenNoteOnIndex];
 
                 activePianoNotes.push({
                     midi: forgivenNoteOn.midi,
                     createdAt: {
                         seconds: forgivenNoteOn.createdAt.seconds,
                         tick: forgivenNoteOn.createdAt.tick,
-                    }
-                })
+                    },
+                });
             }
 
-            const cleanedEventBuffer = bufferEvent.filter((el, index) => !indexesToRemove.includes(index));
-            // console.log(`bufferEvent ${bufferEvent.length} cleanedEventBuffer ${cleanedEventBuffer.length} activePianoNotes ${activePianoNotes.length}`);
-            // iterate over the active notes. If the note exists in the bufferEvent as an "on" event, then we have a note on
-            // if the note does not exist in the bufferEvent, then we have a continuation of the note (we assume it was activated in a previous tick)
+            const cleanedEventBuffer = bufferEvent.filter(
+                (el, index) => !indexesToRemove.includes(index));
+            // iterate over the active notes. If the note exists in the bufferEvent
+            //  as an "on" event, then we have a note on
+            // if the note does not exist in the bufferEvent, then we have a continuation
+            // of the note (we assume it was activated in a previous tick)
             for (let i = 0; i < activePianoNotes.length; i++) {
                 const midi = activePianoNotes[i].midi;
-                const noteOnEvent = cleanedEventBuffer.find(elem => elem.type === vm.noteType.NOTE_ON && elem.midi === midi);
+                const noteOnEvent = cleanedEventBuffer.find(
+                    (elem) => elem.type === vm.noteType.NOTE_ON && elem.midi === midi);
                 if (noteOnEvent) {
-                    // currentQuantizedEvents.push({
-                    //   type: "on",
-                    //   midi: midi,
-                    // })
-                    currentQuantizedEvents.push(noteOnEvent)
-                }
-                else {
+                    currentQuantizedEvents.push(noteOnEvent);
+                } else {
                     const noteHoldEvent = new NoteEvent();
                     noteHoldEvent.type = vm.noteType.NOTE_HOLD;
                     noteHoldEvent.player = vm.playerType.HUMAN;
@@ -726,9 +743,11 @@ export default {
                 }
             }
             // console.log("currentQuantEvents", currentQuantizedEvents);
-            // TODO : it seems I ignore rests. If no active notes, then currentQuantizedEvents will be empty
+            // TODO : it seems I ignore rests. If no active notes,
+            // then currentQuantizedEvents will be empty
 
-            // now iterate over the bufferEvent and find all the noteOff notes and push them to currentQuantizedEvents as off events
+            // now iterate over the bufferEvent and find all the noteOff notes
+            // and push them to currentQuantizedEvents as off events
             // for (let i = 0; i < cleanedEventBuffer.length; i++) {
             //   const elem = cleanedEventBuffer[i];
             //   if (elem.type === "noteOff") {
@@ -739,29 +758,30 @@ export default {
             //   }
             // }
 
-            // Due to the order we were pushing the events, the first elements of currentQuantizedEvents are the "on", 
-            // then the "hold" and finally the "off". Also because activeNotes are already sorted by their "noteOn" timestamp
-            // the events in currentQuantizedEvents are also sorted by their original "noteOn" timestamp
+            // Due to the order we were pushing the events, the first elements
+            // of currentQuantizedEvents are the "on",
+            // then the "hold" and finally the "off". Also because activeNotes
+            // are already sorted by their "noteOn" timestamp
+            // the events in currentQuantizedEvents are also sorted by their original
+            // "noteOn" timestamp
 
-            // Now if we want we can constraint the polyphony. If polyphony = 3, then we can only have 3 notes on at the same time
-            // the way to do that is to keep at most the first 3 "on" or "hold" events in currentQuantizedEvents and remove the rest "on" and "hold"
+            // Now if we want we can constraint the polyphony. If polyphony = 3,
+            // then we can only have 3 notes on at the same time
+            // the way to do that is to keep at most the first 3 "on" or "hold"
+            // events in currentQuantizedEvents and remove the rest "on" and "hold"
 
             // TODO : FOR NOW WE DON"T INCLUDE NOTE_OFF EVENTS IN THE QUANTIZED DATA.
             let constrainedCurrentQuantizedEvents = [];
-            let onHoldEvents = currentQuantizedEvents.filter(elem => elem.type === vm.noteType.NOTE_ON || elem.type === vm.noteType.NOTE_HOLD);
+            const onHoldEvents = currentQuantizedEvents.filter(
+                (elem) => elem.type === vm.noteType.NOTE_ON || elem.type === vm.noteType.NOTE_HOLD);
             // let offEvents = currentQuantizedEvents.filter(elem => elem.type === "off");
             if (onHoldEvents.length > vm.config.noteModeSettings.gridBased.polyphony.input) {
                 // let onHoldEventsToRemove = onHoldEvents.slice(polyphony);
-                let onHoldEventsToKeep = onHoldEvents.slice(0, vm.config.noteModeSettings.gridBased.polyphony.input);
-                // let offEventsToAdd = onHoldEventsToRemove.map(elem => {
-                //   return {
-                //     type: "off",
-                //     midi: elem.midi,
-                //   }
-                // })
-                constrainedCurrentQuantizedEvents = [...onHoldEventsToKeep];//, ...offEventsToAdd, ...offEvents];
-            }
-            else {
+                const onHoldEventsToKeep = onHoldEvents.
+                    slice(0, vm.config.noteModeSettings.gridBased.polyphony.input);
+                constrainedCurrentQuantizedEvents = [...onHoldEventsToKeep];
+                // , ...offEventsToAdd, ...offEvents];
+            } else {
                 constrainedCurrentQuantizedEvents = [...currentQuantizedEvents];
             }
             // console.log("currentQuantEvents", currentQuantizedEvents);
@@ -769,29 +789,30 @@ export default {
             // constrainedCurrentQuantizedEvents.forEach(elem => {
             //     console.log("type ", elem.type);
             // })
-            this.$store.dispatch("storeHumanQuantizedInput", constrainedCurrentQuantizedEvents);
+            this.$store.dispatch('storeHumanQuantizedInput', constrainedCurrentQuantizedEvents);
 
-            this.$store.commit("clearContinuousBuffers");
+            this.$store.commit('clearContinuousBuffers');
             // TODO : for the future, keep a reference to the active notes of the previous tick
-            // TODO : for more accuracy keep also a reference to the previous tick's quantized notes and constrained quantized notes
+            // TODO : for more accuracy keep also a reference to the previous tick's quantized
+            // notes and constrained quantized notes
             // TODO : and modify the logic accordingly.
         },
 
         agentParameterObserver() {
-            let newParameterAgent = { index: null, value: null };
+            const newParameterAgent = {index: null, value: null};
             if (this.paramReader.dequeue_change(newParameterAgent)) {
                 this.dataForMonitoring[newParameterAgent.index] = newParameterAgent.value;
             }
         },
 
         triggerAgentSamplerSync() {
-            var vm = this;
+            const vm = this;
             const agentNotesToBePlayed = this.$store.getters.popAgentNotesToBePlayedAt(
-                this.$store.getters.getGlobalTickDelayed
+                this.$store.getters.getGlobalTickDelayed,
             );
             // This guy here should update the lastNoteAI for the score to get
             // naively choose the first note only. ScoreUI only supports monophonic
-            this.$store.dispatch("updateLastAgentNote", agentNotesToBePlayed);
+            this.$store.dispatch('updateLastAgentNote', agentNotesToBePlayed);
             if (agentNotesToBePlayed.length > 0) {
                 agentNotesToBePlayed.forEach((noteEvent) => {
                     vm.processAgentNoteEvent(noteEvent);
@@ -803,40 +824,43 @@ export default {
 
         async agentCallback(e) {
             const vm = this;
-            let hookType = parseInt(e.data.hookType);
-            let message = e.data.message;
+            const hookType = parseInt(e.data.hookType);
+            const message = e.data.message;
             // TODO move CLOCK_TIME and INFERENCE_TIME in the case/switch
             // TODO that will make those 2 optional in the agent hook.
             if (hookType == vm.agentHookType.CLOCK_EVENT) {
                 // Look for the CLOCK_TIME and INFERENCE_TIME messages
-                // The agent should always include two messages of type CLOCK_TIME and INFERENCE_TIME
+                // The agent should always include two messages of type
+                // CLOCK_TIME and INFERENCE_TIME
                 // when posting from the CLOCK_EVENT hook (processClockEvent())
-                let agentPredictionTick = e.data.message[vm.messageType.CLOCK_TIME];
-                if ((agentPredictionTick !== this.$store.getters.getLocalTickDelayed) && (this.$store.getters.getGlobalTick > 2)) {
+                const agentPredictionTick = e.data.message[vm.messageType.CLOCK_TIME];
+                if ((agentPredictionTick !== this.$store.getters.getLocalTickDelayed)&&
+                        (this.$store.getters.getGlobalTick > 2)) {
                     this.$toasted.show(
-                        "Network tick misalignment: expecting " +
-                        this.$store.getters.getLocalTickDelayed +
-                        ", got " +
-                        agentPredictionTick
+                        'Network tick misalignment: expecting ' +
+                                    this.$store.getters.getLocalTickDelayed +
+                                    ', got ' +
+                                    agentPredictionTick,
                     );
                     this.misalignErrCount += 1;
                 }
-                let agentInferenceTime = e.data.message[vm.messageType.INFERENCE_TIME];
+                const agentInferenceTime = e.data.message[vm.messageType.INFERENCE_TIME];
                 vm.modelInferenceTimes.push(agentInferenceTime);
                 // console.log("just pushed ", agentInferenceTime, " to modelInferenceTimes")
             }
 
-            for (let messageTypeStr in message) {
-                let messageValue = message[messageTypeStr];
-                let messageType = parseInt(messageTypeStr);
+            for (const messageTypeStr in message) {
+                if (Object.hasOwn(message, messageTypeStr)) {
+                    const messageValue = message[messageTypeStr];
+                    const messageType = parseInt(messageTypeStr);
 
-                switch (messageType) {
+                    switch (messageType) {
                     case vm.messageType.STATUS:
                         if (messageValue == vm.statusType.SUCCESS) {
-                            vm.$refs.entryBtn.classList.add("fade-in");
-                            vm.$refs.entryBtn.style.visibility = "visible";
+                            vm.$refs.entryBtn.classList.add('fade-in');
+                            vm.$refs.entryBtn.style.visibility = 'visible';
                             vm.modelLoadTime = Date.now() - vm.modelLoadTime;
-                            console.log("success");
+                            console.log('success');
                         }
                         break;
                     case vm.messageType.NOTE_LIST: {
@@ -846,10 +870,10 @@ export default {
 
                             // The noteEvents received from the agent are serialized
                             // we need to deserialize them before using them
-                            let noteEvent = NoteEvent.fromPlain(noteEventPlain);
+                            const noteEvent = NoteEvent.fromPlain(noteEventPlain);
                             if (noteEvent.playAfter.tick > 0) {
                                 // console.log("NAI NAI NAI NIA NIA NIA ");
-                                this.$store.dispatch("storeAgentQuantizedOutput", noteEvent);
+                                this.$store.dispatch('storeAgentQuantizedOutput', noteEvent);
                             } else {
                                 vm.processAgentNoteEvent(noteEvent);
                             }
@@ -857,11 +881,13 @@ export default {
                         break;
                     }
                     case vm.messageType.CHROMA_VECTOR:
-                        if (this.audioAndChroma)
+                        if (this.audioAndChroma) {
                             this.$root.$refs.chromaChart.updateChromaData(messageValue);
-                        // this.$root.$refs.vectorBar.updateVectorData(messageValue);
-                        else
-                            console.warn("The Agent generates chroma vectors but the Chroma Chart is not enabled in the config file");
+                            // this.$root.$refs.vectorBar.updateVectorData(messageValue);
+                        } else {
+                            console.warn('The Agent generates chroma vectors but the Chroma' +
+                            'Chart is not enabled in the config file');
+                        }
                         break;
                     case vm.messageType.LABEL:
                         this.textBoxText = messageValue;
@@ -870,76 +896,80 @@ export default {
                         if (hookType == vm.agentHookType.INIT_AGENT) {
                             const agentStatus = vm.$refs.agentStatus;
                             agentStatus.innerHTML = messageValue;
-                        }
-                        else {
+                        } else {
                             this.textBoxText = messageValue;
                         }
                         break;
-                    // TODO : 
+                        // TODO :
                     case vm.messageType.CLOCK_TIME:
                         break;
                     case vm.messageType.INFERENCE_TIME:
                         break;
                     default:
-                        console.log("Unknown message type: ", messageType);
+                        console.log('Unknown message type: ', messageType);
                         break;
+                    }
                 }
             }
         },
 
         uiNoteOnAgent(noteEvent) {
-            let vm = this;
+            const vm = this;
             // set a timeout to call keyDown based on noteEvent.timestamp.seconds
             // TODO : the agent should do that
-            let noteName = Midi.midiToNoteName(noteEvent.midi, { sharps: true });
-            let whiteKey = noteName.includes('#') ? false : true;
+            const noteName = Midi.midiToNoteName(noteEvent.midi, {sharps: true});
+            const whiteKey = noteName.includes('#') ? false : true;
             // Check if the current on-screen key is within the current screen view
-            // if (this.$root.$refs.keyboard.$refs[noteEvent.name] ) is null then the key is not on screen
-            let keyOnScreenRange = this.$root.$refs.keyboard.$refs[noteName] ? true : false;
+            // if (this.$root.$refs.keyboard.$refs[noteEvent.name] ) is
+            // null then the key is not on screen
+            const keyOnScreenRange = this.$root.$refs.keyboard.$refs[noteName] ? true : false;
             vm.timeout_IDS_kill.push(setTimeout(() => {
                 // if (this.config.gui.pianoRoll.status)
-                if (this.config.gui.pianoRoll.status && this.config.gui.pianoRoll.agent)
+                if (this.config.gui.pianoRoll.status && this.config.gui.pianoRoll.agent) {
                     this.$root.$refs.pianoRoll.keyDown(noteEvent);
+                }
                 if (keyOnScreenRange) {
                     if (whiteKey) {
-                        this.$root.$refs.keyboard.$refs[noteName][0].classList.add('active-white-key-agent');
+                        this.$root.$refs.keyboard.$refs[noteName][0].
+                            classList.add('active-white-key-agent');
                     } else {
-                        this.$root.$refs.keyboard.$refs[noteName][0].classList.add('active-black-key-agent');
+                        this.$root.$refs.keyboard.$refs[noteName][0].
+                            classList.add('active-black-key-agent');
                     }
                 }
-            }, noteEvent.playAfter.seconds * 1000)
+            }, noteEvent.playAfter.seconds * 1000),
             );
         },
 
         // TODO : refactor this one. choose which option
         // to use and remove the other one
         processAgentNoteEvent(noteEvent) {
-            let vm = this;
-            // The noteEvents that arrive here, have already waited for playAfter.tick ticks (if any)
-            // so we can only care about playAfter.seconds here. 
+            const vm = this;
+            // The noteEvents that arrive here, have already
+            //  waited for playAfter.tick ticks (if any)
+            // so we can only care about playAfter.seconds here.
             if (noteEvent.type === vm.noteType.NOTE_ON) {
-
-                // Here we check if this note has a defined duration. 
+                // Here we check if this note has a defined duration.
                 if (noteEvent.duration != null) {
-                    // Two ways to dealwith notes that have predefined duration.
-                    // 1) Create a matching NOTE_OFF event and schedule it to be played after 'duration'
-                    // 2) Use Tone.js triggerAttackRelease() api, which takes care of the NOTE_OFF event
+                // Two ways to dealwith notes that have predefined duration.
+                // 1) Create a matching NOTE_OFF event and schedule it to be played after 'duration'
+                // 2) Use Tone.js triggerAttackRelease() api, which takes care of the NOTE_OFF event
 
                     if (noteEvent.duration.tick > 0) {
                         // ---------------- Second OPTION ----------------
                         if (vm.config.custom.useTriggerRelease) {
-                            this.$store.dispatch("samplerOnOff", noteEvent);
-                            // console.log("option2  ")
+                            this.$store.dispatch('samplerOnOff', noteEvent);
+                        // console.log("option2  ")
                         } else {
-                            this.$store.dispatch("samplerOn", noteEvent);
-                            // console.log("option1 ")
+                            this.$store.dispatch('samplerOn', noteEvent);
+                        // console.log("option1 ")
                         }
                         // else {
-                        // If its duration is specified in ticks, then 
+                        // If its duration is specified in ticks, then
                         // create a noteEvent with the same properties but type NOTE_OFF
                         // and modify its playAfter accordingly. Also set its duration to null
                         // so that it doesn't get processed again in this if statement
-                        let noteEventOff = NoteEvent.copy(noteEvent);
+                        const noteEventOff = NoteEvent.copy(noteEvent);
                         noteEventOff.type = vm.noteType.NOTE_OFF;
                         noteEventOff.playAfter.tick = noteEvent.duration.tick;
                         noteEventOff.playAfter.seconds += noteEvent.duration.seconds;
@@ -949,94 +979,97 @@ export default {
                         if (vm.config.custom.useTriggerRelease) {
                             noteEventOff.custom = true;
                         }
-                        this.$store.dispatch("storeAgentQuantizedOutput", noteEventOff);
+                        this.$store.dispatch('storeAgentQuantizedOutput', noteEventOff);
                         // }
-
                     } else if (noteEvent.duration.tick == 0) {
                         if (noteEvent.duration.seconds > 0) {
                             if (vm.config.custom.useTriggerRelease) {
                                 // console.log("about to send samplerOnOff ", noteEvent.midi);
-                                this.$store.dispatch("samplerOnOff", noteEvent);
+                                this.$store.dispatch('samplerOnOff', noteEvent);
                                 vm.timeout_IDS_live.push(setTimeout(() => {
                                     this.uiNoteOffAgent(noteEvent);
-                                }, noteEvent.duration.seconds * 1000)
+                                }, noteEvent.duration.seconds * 1000),
                                 );
 
                                 // console.log("option2  ")
                             } else {
-                                this.$store.dispatch("samplerOn", noteEvent);
-                                console.log("option1 ")
+                                this.$store.dispatch('samplerOn', noteEvent);
+                                console.log('option1 ');
+                                // TODO
                                 // if I want option1 when duration is 0ticks and >0seconds
                                 // I need to create a new noteOff event, and send it after >0seconds
-                                // to a new method  that will call samplerOff and uiNoteOffAgent. TODO
+                                // to a new method  that will call samplerOff and uiNoteOffAgent.
                                 // For now, I just use option2
                                 vm.timeout_IDS_live.push(setTimeout(() => {
                                     this.uiNoteOffAgent(noteEvent);
-                                    this.$store.dispatch("samplerOff", noteEvent)
-                                }, noteEvent.duration.seconds * 1000)
+                                    this.$store.dispatch('samplerOff', noteEvent);
+                                }, noteEvent.duration.seconds * 1000),
                                 );
                             }
                         } else {
-                            // throw error, agent generated note with duration 0ticks and 0seconds
-                            console.error("Agent generated note with duration 0 ticks and 0 seconds")
+                        // throw error, agent generated note with duration 0ticks and 0seconds
+                            console.error('Agent generated note with' +
+                            ' duration 0 ticks and 0 seconds');
                         }
                     }
-
-                }
-                else {
-                    this.$store.dispatch("samplerOn", noteEvent);
+                } else {
+                    this.$store.dispatch('samplerOn', noteEvent);
                 }
                 // this.$store.dispatch("samplerOn", noteEvent);
                 this.uiNoteOnAgent(noteEvent);
             } else if (noteEvent.type === vm.noteType.NOTE_OFF) {
                 if (!noteEvent.hasOwnProperty('custom')) {
-                    this.$store.dispatch("samplerOff", noteEvent);
+                    this.$store.dispatch('samplerOff', noteEvent);
                 }
                 // console.log("about to call uiNoteOffAgent");
                 this.uiNoteOffAgent(noteEvent);
             } else if (noteEvent.type === vm.noteType.NOTE_HOLD) {
-                console.log("HOLD event received from agent");
+                console.log('HOLD event received from agent');
             }
         },
 
         uiNoteOffAgent(noteEvent) {
-            let vm = this;
+            const vm = this;
             // console.log("uiNoteOffAgent")
             // TODO : the agent should do that
-            let noteName = Midi.midiToNoteName(noteEvent.midi, { sharps: true });
+            const noteName = Midi.midiToNoteName(noteEvent.midi, {sharps: true});
             // If '#' in noteName then whiteKey is false else true
-            let whiteKey = noteName.includes('#') ? false : true;
+            const whiteKey = noteName.includes('#') ? false : true;
             // Check if the current on-screen key is within the current screen view
-            // if (this.$root.$refs.keyboard.$refs[noteEvent.name] ) is null then the key is not on screen
-            let keyOnScreenRange = this.$root.$refs.keyboard.$refs[noteName] ? true : false;
+            // if (this.$root.$refs.keyboard.$refs[noteEvent.name] )
+            // is null then the key is not on screen
+            const keyOnScreenRange = this.$root.$refs.keyboard.$refs[noteName] ? true : false;
             // set a timeout to call keyUp based on noteEvent.playAfter.seconds
             vm.timeout_IDS_live.push(setTimeout(() => {
                 if (keyOnScreenRange) {
                     if (whiteKey) {
                         // console.log("released white key ", noteName)
-                        this.$root.$refs.keyboard.$refs[noteName][0].classList.remove('active-white-key-agent');
+                        this.$root.$refs.keyboard.$refs[noteName][0].
+                            classList.remove('active-white-key-agent');
                     } else {
                         // console.log("released black key ", noteName)
-                        this.$root.$refs.keyboard.$refs[noteName][0].classList.remove('active-black-key-agent');
+                        this.$root.$refs.keyboard.$refs[noteName][0].
+                            classList.remove('active-black-key-agent');
                     }
                 }
                 // if (this.config.gui.pianoRoll.status)
-                if (this.config.gui.pianoRoll.status && this.config.gui.pianoRoll.agent)
+                if (this.config.gui.pianoRoll.status && this.config.gui.pianoRoll.agent) {
                     this.$root.$refs.pianoRoll.keyUp(noteEvent);
-            }, noteEvent.playAfter.seconds * 1000)
+                }
+            }, noteEvent.playAfter.seconds * 1000),
             );
         },
 
         /*
-        * Web MIDI
-        */
+            * Web MIDI
+            */
         onEnabled() {
             const vm = this;
             if (WebMidi.inputs.length < 1) {
                 vm.activeDevices = [];
             } else {
                 WebMidi.inputs.forEach((device) => {
-                    vm.activeDevices.push({ id: device.id, name: device.name });
+                    vm.activeDevices.push({id: device.id, name: device.name});
                 });
                 this.selectedMIDIDevice = this.activeDevices[0].id;
                 this.messageListener();
@@ -1057,7 +1090,7 @@ export default {
         messageListener() {
             const vm = this;
             const inputDevice = WebMidi.getInputById(this.selectedMIDIDevice);
-            inputDevice.addListener("noteon", (message) => {
+            inputDevice.addListener('noteon', (message) => {
                 const newNoteEvent = new NoteEvent();
                 newNoteEvent.player = vm.playerType.HUMAN;
                 newNoteEvent.instrument = vm.instrumentType.PIANO;
@@ -1073,14 +1106,14 @@ export default {
                 };
                 newNoteEvent.playAfter = {
                     seconds: 0,
-                    tick: 0
+                    tick: 0,
                 };
                 newNoteEvent.duration = null;
 
-                vm.processUserNoteEvent(newNoteEvent)
+                vm.processUserNoteEvent(newNoteEvent);
             });
 
-            inputDevice.addListener("noteoff", (message) => {
+            inputDevice.addListener('noteoff', (message) => {
                 const newNoteEvent = new NoteEvent();
                 newNoteEvent.player = vm.playerType.HUMAN;
                 newNoteEvent.instrument = vm.instrumentType.PIANO;
@@ -1096,28 +1129,26 @@ export default {
                 };
                 newNoteEvent.playAfter = {
                     seconds: 0,
-                    tick: 0
+                    tick: 0,
                 };
                 newNoteEvent.duration = null;
 
-                vm.processUserNoteEvent(newNoteEvent)
+                vm.processUserNoteEvent(newNoteEvent);
             });
         },
 
         startRecording() {
-            // this.recorderWorkletNode.parameters.get('recordingStatus').setValueAtTime(1, this.audioContext.currentTime);
-            this.$store.commit("startUnMute");
+            this.$store.commit('startUnMute');
         },
 
         stopRecording() {
-            // this.recorderWorkletNode.parameters.get('recordingStatus').setValueAtTime(0, this.audioContext.currentTime);
-            this.$store.commit("stopMute");
+            this.$store.commit('stopMute');
             // this is a nice hack
-            let id = setTimeout(function () { }, 0);
+            let id = setTimeout(function() { }, 0);
             while (id--) {
                 // check if id is in timeout_IDS_live. if not, then clear it
                 if (this.timeout_IDS_kill.includes(id)) {
-                    // console.log("deleting timeout id " + id);
+                // console.log("deleting timeout id " + id);
                     clearTimeout(id); // will do nothing if no timeout with id is present
                 }
             }
@@ -1131,15 +1162,15 @@ export default {
         */
         entryProgram() {
             const vm = this;
-            vm.$refs.mainLoadingScreen.classList.add("fade-out");
-            vm.$refs.mainLoadingScreen.style.display = "none";
-            vm.$refs.mainContent.style.display = "block";
-            vm.$modal.show("introModal");
-            console.log("TONE ENTRY ", Tone.now());
+            vm.$refs.mainLoadingScreen.classList.add('fade-out');
+            vm.$refs.mainLoadingScreen.style.display = 'none';
+            vm.$refs.mainContent.style.display = 'block';
+            vm.$modal.show('introModal');
+            console.log('TONE ENTRY ', Tone.now());
         },
 
         showSettingsModal() {
-            this.$modal.show("settingsModal");
+            this.$modal.show('settingsModal');
         },
 
         toggleMonitor() {
@@ -1159,58 +1190,59 @@ export default {
             this.keyboardoctaveEnd -= 1;
         },
 
-        
+
         zoomInOct() {
-            // ugly code to prevent the user from zooming in too much
-            console.log("range before was ", this.keyboardoctaveEnd - this.keyboardoctaveStart)
-            if (this.keyboardoctaveEnd - this.keyboardoctaveStart <= 1){
-                console.log("MIN was reched before")
+        // ugly code to prevent the user from zooming in too much
+            console.log('range before was ', this.keyboardoctaveEnd - this.keyboardoctaveStart);
+            if (this.keyboardoctaveEnd - this.keyboardoctaveStart <= 1) {
+                console.log('MIN was reched before');
                 this.keyboardMinRangeDisplayed = true;
             } else {
-                if (this.keyboardoctaveEnd - this.keyboardoctaveStart == 2){
+                if (this.keyboardoctaveEnd - this.keyboardoctaveStart == 2) {
                     this.keyboardoctaveStart = Math.max(0, this.keyboardoctaveStart + 1);
-                } else{
+                } else {
                     this.keyboardoctaveStart = Math.max(0, this.keyboardoctaveStart + 1);
                     this.keyboardoctaveEnd = Math.min(8, this.keyboardoctaveEnd - 1);
                 }
-                console.log("this.keyboardoctaveEnd - this.keyboardoctaveStart != 1")
-                if (this.keyboardoctaveEnd - this.keyboardoctaveStart <= 1){
-                    console.log("MIN range reached")
+                console.log('this.keyboardoctaveEnd - this.keyboardoctaveStart != 1');
+                if (this.keyboardoctaveEnd - this.keyboardoctaveStart <= 1) {
+                    console.log('MIN range reached');
                     this.keyboardMinRangeDisplayed = true;
                 } else {
                     this.keyboardMinRangeDisplayed = false;
                 }
-            }   
+            }
             this.keyboardMaxRangeDisplayed = false;
         },
 
         zoomOutOct() {
-            console.log("range before was ", this.keyboardoctaveEnd - this.keyboardoctaveStart)
-            if (this.keyboardoctaveEnd - this.keyboardoctaveStart >= 8){
-                console.log("MAX was reched before")
+            console.log('range before was ', this.keyboardoctaveEnd - this.keyboardoctaveStart);
+            if (this.keyboardoctaveEnd - this.keyboardoctaveStart >= 8) {
+                console.log('MAX was reched before');
                 this.keyboardMaxRangeDisplayed = true;
             } else {
                 this.keyboardoctaveStart = Math.max(0, this.keyboardoctaveStart - 1);
                 this.keyboardoctaveEnd = Math.min(8, this.keyboardoctaveEnd + 1);
-                console.log("this.keyboardoctaveEnd - this.keyboardoctaveStart != 8")
-                if (this.keyboardoctaveEnd - this.keyboardoctaveStart >= 8){
-                    console.log("MAX range reached")
+                console.log('this.keyboardoctaveEnd - this.keyboardoctaveStart != 8');
+                if (this.keyboardoctaveEnd - this.keyboardoctaveStart >= 8) {
+                    console.log('MAX range reached');
                     this.keyboardMaxRangeDisplayed = true;
                 } else {
                     this.keyboardMaxRangeDisplayed = false;
                 }
-                
-            } 
-            this.keyboardMinRangeDisplayed = false; 
+            }
+            this.keyboardMinRangeDisplayed = false;
         },
 
         modalCallback() {
-            this.$store.commit("changeModalStatus");
+            this.$store.commit('changeModalStatus');
         },
 
         calculateMaxBPM() {
             const vm = this;
-            const dt = vm.modelInferenceTimes.sort(function (a, b) { return a - b })[Math.floor(vm.modelInferenceTimes.length * 0.95)];
+            const dt = vm.modelInferenceTimes.sort(function(a, b) {
+                return a - b;
+            })[Math.floor(vm.modelInferenceTimes.length * 0.95)];
             vm.maxBPM = Math.round(1000 * 60 / dt / vm.$store.getters.getTicksPerBeat);
         },
 
@@ -1218,44 +1250,47 @@ export default {
             console.log(value);
             this.localBPM = value;
             let revertToDefault = false;
-            if (this.localSyncClockStatus)
+            if (this.localSyncClockStatus) {
                 this.stopClock();
-                revertToDefault = true;
-            this.clockWorker.postMessage({ action: "setBPM", bpm: value });
-            if (revertToDefault)
+            }
+            revertToDefault = true;
+            this.clockWorker.postMessage({action: 'setBPM', bpm: value});
+            if (revertToDefault) {
                 this.startClock();
+            }
         },
 
         buttonAction(buttonId) {
             // use paramWriter to write a parameter to the agent
             const buttonPropertyName = `BUTTON_${buttonId}`;
             if (this.paramWriter != null &&
-                !this.paramWriter.enqueue_change(this.uiParameterType[buttonPropertyName], 1.0)) {
-                console.warn("Couldn't enqueue.");
+                        !this.paramWriter.enqueue_change(
+                            this.uiParameterType[buttonPropertyName], 1.0)) {
+                console.warn('Couldn\'t enqueue.');
             }
         },
 
         handleMixerUpdate(event) {
-            this.$store.commit("handleMixerUpdate", event);
+            this.$store.commit('handleMixerUpdate', event);
         },
 
         manuallyUpdateMixer(event) {
             this.mixer_data = event;
-            console.log("Main got mixer update", event);
+            console.log('Main got mixer update', event);
         },
 
         manuallyUpdateData() {
-            let vm = this;
+            const vm = this;
             for (const [player, playerData] of Object.entries(vm.mixer_data)) {
-                console.log("player", player);
+                console.log('player', player);
                 for (const [instId, instData] of Object.entries(playerData.instruments)) {
-                    console.log("instId", instId);
-                    let changeEvent = {
+                    console.log('instId', instId);
+                    const changeEvent = {
                         playerId: player,
                         instrumentId: instId,
                         what: 'volume',
-                        value: instData.volume
-                    }
+                        value: instData.volume,
+                    };
                     vm.handleMixerUpdate(changeEvent);
                 };
             }
@@ -1264,34 +1299,34 @@ export default {
         loadConfigSync() {
             // Read about sync and async requests here:
             // https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/Synchronous_and_Asynchronous_Requests
-            // In our case, we need to load the config asap, since 
+            // In our case, we need to load the config asap, since
             // the config contains also info to generate the UI.
             // If not, the app will crash anyway, so it's worth waiting for it.
 
             // First, load all the configs files and merge them into one
-            let configFilesURL = this.configFiles.map((file) => {
+            const configFilesURL = this.configFiles.map((file) => {
                 return `/agents/${this.agentName}/${file}`;
             });
             // get all files using xhr
-            let xhrs = configFilesURL.map((url) => {
+            const xhrs = configFilesURL.map((url) => {
                 const xhr = new XMLHttpRequest();
-                xhr.open("GET", url, false); // Set async to false
+                xhr.open('GET', url, false); // Set async to false
                 xhr.send();
                 return xhr;
             });
             // go over the xhrs, check which are 200, and concat them
             // for the rest of the files, throw an error
-            let config = "";
+            let config = '';
             xhrs.forEach((xhr) => {
                 if (xhr.status === 200) {
-                    config += xhr.responseText + "\n";
+                    config += xhr.responseText + '\n';
                 } else {
                     throw new Error(`Failed to fetch config file: ${xhr.status}`);
                 }
             });
 
-            if (config === "") {
-                throw new Error("Failed to fetch config file: empty");
+            if (config === '') {
+                throw new Error('Failed to fetch config file: empty');
             } else {
                 this.config = yaml.load(config);
             }
@@ -1303,30 +1338,37 @@ export default {
         // so that we can watch() the computedSwitches while
         // havin access to the old switches array.
         // https://github.com/vuejs/vue/issues/2164
-        computedSwitches: function () {
-            return this.switches.map(a => { return { ...a } })
+        computedSwitches: function() {
+            return this.switches.map((a) => {
+                return {...a};
+            });
         },
 
-        computedSliders: function () {
-            return this.sliders.map(a => { return { ...a } })
+        computedSliders: function() {
+            return this.sliders.map((a) => {
+                return {...a};
+            });
         },
-        audioAndMeter: function () {
+        audioAndMeter: function() {
             return this.config.gui.audioMeter.status && this.config.interactionMode.audioMode;
         },
-        audioAndChroma: function () {
+        audioAndChroma: function() {
             return this.config.gui.chromaChart.status && this.config.interactionMode.audioMode;
         },
-        showMonitor: function () {
-            if (this.config.gui.monitor)
-                if (this.config.gui.monitor.status)
+        showMonitor: function() {
+            if (this.config.gui.monitor) {
+                if (this.config.gui.monitor.status) {
                     return true;
+                }
+            }
             return false;
-        }
+        },
     },
 
     watch: {
         // screenWidth: {
-        //     // At every screenWidth data change, this would automatically change the keyboard's octave number.
+        //     // At every screenWidth data change, this would
+        //     // automatically change the keyboard's octave number.
         //     immediate: true,
         //     handler(newValue) {
         //         let octaves;
@@ -1355,7 +1397,7 @@ export default {
             handler(newValue) {
                 if (newValue == 10) {
                     this.$toasted.show(
-                        "Your local machine cannot run inference at this speed. Try lowering the BPM."
+                        'Your machine cannot run inference at this speed. Try lowering the BPM.',
                     );
                 }
             },
@@ -1372,32 +1414,32 @@ export default {
         localBPM: {
             immediate: true,
             handler(newValue) {
-                console.log("localBPM changed to ", newValue);
-                this.$store.commit("setBPM", newValue);
+                console.log('localBPM changed to ', newValue);
+                this.$store.commit('setBPM', newValue);
             },
         },
 
         computedSwitches: {
-            // immediate: true,
+        // immediate: true,
             deep: true,
             handler(newStates, oldStates) {
                 newStates.forEach((newState, index) => {
-                    let oldStatus = oldStates.length == 0 ? null : oldStates[index].status;
+                    const oldStatus = oldStates.length == 0 ? null : oldStates[index].status;
                     if (newState.status !== oldStatus) {
-                        let floatStatus = newState.status ? 1.0 : 0.0;
+                        const floatStatus = newState.status ? 1.0 : 0.0;
                         const switchPropertyName = `SWITCH_${index + 1}`;
                         if (
                             this.paramWriter != null &&
-                            !this.paramWriter.enqueue_change(
-                                this.uiParameterType[switchPropertyName],
-                                floatStatus
-                            )
+                                            !this.paramWriter.enqueue_change(
+                                                this.uiParameterType[switchPropertyName],
+                                                floatStatus,
+                                            )
                         ) {
-                            console.warn("Couldn't enqueue.");
+                            console.warn('Couldn\'t enqueue.');
                         }
                     }
                 });
-            }
+            },
         },
 
         computedSliders: {
@@ -1405,26 +1447,26 @@ export default {
             deep: true,
             handler(newStates, oldStates) {
                 newStates.forEach((newState, index) => {
-                    let oldStatus = oldStates.length == 0 ? null : oldStates[index].status;
+                    const oldStatus = oldStates.length == 0 ? null : oldStates[index].status;
                     if (newState.value !== oldStatus) {
                         // let floatStatus = newState.status ? 1.0 : 0.0;
                         const sliderPropertyName = `SLIDER_${index + 1}`;
-                        console.log("detected ", sliderPropertyName, " change to ", newState.value);
+                        console.log('detected ', sliderPropertyName, ' change to ', newState.value);
                         if (
                             this.paramWriter != null &&
-                            !this.paramWriter.enqueue_change(
-                                this.uiParameterType[sliderPropertyName],
-                                newState.value
-                            )
+                                            !this.paramWriter.enqueue_change(
+                                                this.uiParameterType[sliderPropertyName],
+                                                newState.value,
+                                            )
                         ) {
-                            console.warn("Couldn't enqueue.");
+                            onsole.warn('Couldn\'t enqueue.');
                         }
                     }
                 });
-            }
+            },
         },
-    }
-}
+    },
+};
 
 </script>
 
@@ -1532,7 +1574,7 @@ export default {
                     <i class="material-symbols-outlined">monitoring</i>
                 </md-button>
             </div>
-            
+
             <md-button v-if="keyboardoctaveEnd !== 8" @click="transposeOctUp" class="md-icon-button md-raised"
                 style="position: absolute; right: 20px; bottom: 100px">
                 <md-icon>arrow_forward</md-icon>
@@ -1577,14 +1619,16 @@ export default {
                                     <div class="settingsDiv">
                                         <p class="settingsOptionTitle">BPM (Max: {{ maxBPM }})</p>
                                         <div style="padding-top: 14px">
-                                            <BPMSlider v-model="localBPM" :min="60" :max="120" @bpmChangeEvent="bpmValueChanged"/> 
+                                            <BPMSlider v-model="localBPM" :min="60" :max="120"
+                                                        @bpmChangeEvent="bpmValueChanged"/>
                                             <!-- @bpmChangeEvent="bpmValueChanged" -->
                                         </div>
                                     </div>
                                 </div>
                             </div>
                         </div>
-                        <div id="MIDISection" style="display:flex;justify-content:start;align-items:center;margin-top:50px">
+                        <div id="MIDISection" style="display:flex;
+                                justify-content:start;align-items:center;margin-top:50px">
                             <p class="settingsSubtitle settingsBorderedTitle" style="padding:0">
                                 <md-icon class="modalIcon" style="font-weight:400">piano</md-icon> <br />
                                 <span style="line-height:36px">MIDI</span>
@@ -1609,7 +1653,8 @@ export default {
                                     <div v-for="sliderItem in sliders" :key="sliderItem.id"
                                         class="md-layout-item md-large-size-25 md-alignment-center"
                                         style="display:flex;align-items:center;justify-content:center;">
-                                        <VerticalSlider v-model="sliderItem.value" :min="sliderItem.min" :max=sliderItem.max
+                                        <VerticalSlider v-model="sliderItem.value"
+                                                        :min="sliderItem.min" :max=sliderItem.max
                                             :label="sliderItem.label" />
                                     </div>
                                 </div>
